+++ cutplace/test_tools.py
@@ -146,22 +146,22 @@ class ToolsTest(unittest.TestCase):
 
 class NumberedTest(unittest.TestCase):
     def testCanDetectNoneNumber(self):
-        self.assertEqual(_tools.numbered(u"123abc"), (None, u"123abc"))
-        self.assertEqual(_tools.numbered(u"01.02.2014"), (None, u"01.02.2014"))
+        self.assertEqual(_tools.numbered(u"123abc"), (None, False, u"123abc"))
+        self.assertEqual(_tools.numbered(u"01.02.2014"), (None, False, u"01.02.2014"))
 
     def testCanDetectInteger(self):
-        self.assertEqual(_tools.numbered("123"), (_tools.NUMBER_INTEGER, 123))
+        self.assertEqual(_tools.numbered("123"), (_tools.NUMBER_INTEGER, False, 123))
 
     def testCanDetectDecimalWithPoint(self):
-        self.assertEqual(_tools.numbered("123.45"), (_tools.NUMBER_DECIMAL_POINT, decimal.Decimal("123.45")))
-        self.assertEqual(_tools.numbered("123,456.78"), (_tools.NUMBER_DECIMAL_POINT, decimal.Decimal("123456.78")))
+        self.assertEqual(_tools.numbered("123.45"), (_tools.NUMBER_DECIMAL_POINT, False, decimal.Decimal("123.45")))
+        self.assertEqual(_tools.numbered("123,456.78"), (_tools.NUMBER_DECIMAL_POINT, True, decimal.Decimal("123456.78")))
 
     def testCanDetectDecimalWithComma(self):
-        actual = _tools.numbered("123,45", decimalDelimiter=",", thousandsDelimiter=".")
-        expected = (_tools.NUMBER_DECIMAL_COMMA, decimal.Decimal("123.45"))
+        actual = _tools.numbered("123,45", decimalSeparator=",", thousandsSeparator=".")
+        expected = (_tools.NUMBER_DECIMAL_COMMA, False, decimal.Decimal("123.45"))
         self.assertEqual(actual, expected)
-        actual = _tools.numbered("123.456,78", decimalDelimiter=",", thousandsDelimiter=".")
-        expected = (_tools.NUMBER_DECIMAL_COMMA, decimal.Decimal("123456.78"))
+        actual = _tools.numbered("123.456,78", decimalSeparator=",", thousandsSeparator=".")
+        expected = (_tools.NUMBER_DECIMAL_COMMA, True, decimal.Decimal("123456.78"))
         self.assertEqual(actual, expected)